refactor: move encodedRedirect utility to navigation-utils module
