+++ frontend/utils/navigation-utils.ts
